move riscv64ext to a new file
